move workspace user functionality to its own component
